requirements.txt
flask==3.1.0
flask-cors==5.0.0
Pillow==11.1.0
PyPDF2==3.0.1 